requirements.txt
FastAPI
uvicorn
python-dotenv
google-adk
pydantic
httpx